backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
werkzeug==2.3.7
pyjwt==2.8.0
gunicorn==21.2.0 